requirements.txt
yfinance>=0.2.36
feedparser>=6.0.11
praw>=7.7.1
alpaca-py>=0.20.0
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.20.0
requests>=2.31.0
python-dotenv>=1.0.0
